+++ src/orfimage_int.cpp
@@ -10,9 +10,10 @@ namespace Exiv2::Internal {
     {
     }
 
-    bool OrfHeader::read(const byte* pData, uint32_t size)
+    bool OrfHeader::read(const byte* pData, size_t size)
     {
-        if (size < 8) return false;
+        if (size < 8)
+            return false;
 
         if (pData[0] == 'I' && pData[0] == pData[1]) {
             setByteOrder(littleEndian);
@@ -25,11 +26,12 @@ namespace Exiv2::Internal {
         }
 
         uint16_t sig = getUShort(pData + 2, byteOrder());
-        if (tag() != sig && 0x5352 != sig) return false; // #658: Added 0x5352 "SR" for SP-560UZ
+        if (tag() != sig && 0x5352 != sig)
+            return false; // #658: Added 0x5352 "SR" for SP-560UZ
         sig_ = sig;
         setOffset(getULong(pData + 4, byteOrder()));
         return true;
-    } // OrfHeader::read
+    }
 
     DataBuf OrfHeader::write() const
     {
